requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
Django==3.0
django-filter==2.4.0
future==0.18.2
html5lib==1.1
mysqlclient==2.0.3
Pillow==8.1.0
pycryptodome==3.10.1
pycryptodomex==3.10.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.67
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
xhtml2pdf==0.2.5
